🚚 Rename: 이동 v2 파일들 정리 컴포넌트 폴더에 다 같이 모음
