+++ src/components/BDInfo/BDFreeBoard.tsx
