+++ src/components/BDInfo/BDInfoBoard.tsx
