+++ src/components/BDInfo/BDInfoHome.tsx
