+++ src/components/BDInfo/BDQna.tsx
