--- src/components/BDMag/Magazine.tsx
@@ -1,5 +1,0 @@
-const Magazine = () => {
-  return <div>Magazine</div>;
-};
-
-export default Magazine;
